Re-create the admin client when collecting metrics
The main issue seems to be that the admin client does not recover correctly
when it fails to pick a good broker, so even rewriting how metrics are collected
doesn't really fix the issue.

Now trying to reload the admin client on every collect until we can find out
exactly what is broken and upstream a fix to sarama itself.

Also, JMX isn't an option in this case as kafka brokers do not expose producer/consumer metrics anymore.